Updated covid example
